fix: arena __int128 under-alignment + tk_mul_u16 overflow — re-flip ASan lane to native gate (#327)
* fix: arena __int128 under-alignment + tk_mul_u16 overflow — re-flip ASan lane to native gate

Two pre-existing runtime UBs the native-ASan gate (flipped by #324/#326) surfaced but
the VM gate never reached, both fixed at the root; then re-flip the ASan CI lane back to
the native gate to continuously guard them.

UB 1 — tk_mul_u16 signed-integer-overflow (teko_rt.h:696): the uint16 operands are
integer-PROMOTED to int, so 65535*65535 overflows int (UBSan signed-integer-overflow, hit
by the crypto #tests). Fixed by multiplying in `unsigned`. Audited the sibling
tk_mul_*/tk_add_*/tk_sub_* helpers: only u16 mul overflows via promotion (u8/i8/i16 all fit
int; u32+/i32+ are native-width — no promotion); the TEKO_OVERFLOW_DEBUG twins use
__builtin_*_overflow and are already safe.

UB 2 — systemic arena under-alignment for __int128 (teko_rt.c). tk_region_alloc rounded
size AND bump offset to _Alignof(max_align_t), and the chunk `data` flexible member was
`max_align_t data[]`. On arm64 _Alignof(max_align_t)==8, but ExprKind/TExprKind carry a C
`__int128` needing 16, so every Expr/TExpr through the arena could land 8-byte-aligned →
UBSan alignment errors at hundreds of sites (x86_64 CI never saw it: max_align_t==16 there
already covers __int128). Fix raises the arena's alignment guarantee to
TK_ARENA_ALIGN = max(_Alignof(max_align_t), _Alignof(__int128)) (16 on arm64, still 16 on
x86_64), reconciling EVERY consumer of the old constant:
  - chunk `data` member over-aligned via _Alignas(TK_ARENA_ALIGN) (offset 24→32 on arm64);
  - the chunk-layout _Static_assert now requires TK_ARENA_ALIGN, plus a new assert that the
    struct alignment covers it;
  - chunks allocated through a portable aligned allocator (posix_memalign / _aligned_malloc)
    and released through tk_chunk_free, since malloc only promises _Alignof(max_align_t);
  - tk_region_alloc rounds size AND bump offset to TK_ARENA_ALIGN;
  - the free-list bin rounding (tk_free_take ceil-16 / tk_free_block floor-16) and the
    TEKO_MEM_PARANOID poison path documented against TK_ARENA_ALIGN (bin granularity == 16).

Local arm64 ASan proof (this host is _Alignof(max_align_t)==8, the platform that hits it):
before, the native gate under -fsanitize=address,undefined halt_on_error=0 reported 425
alignment runtime errors + 1 tk_mul_u16 signed-overflow; after the fixes the SAME run is
CLEAN (zero UBSan output), proving no tail behind them. halt_on_error=1 native gate exits 0.

Runtime-seam edit: codegen is untouched, so the emitted C is byte-identical — FIXPOINT
gen1.c == gen2.c == gen3.c (md5 f2f1b663459e2f071ade534f80fff902). diff_vm_native 97/0/1.
TEKO_MEM_PARANOID build exit 0.

sanitizers.yml: re-flip the ASan lane to the NATIVE gate (`./bin/teko . -o gen2-asan`),
restoring the native-flip comment block (#323 ASan caps kept) — the lane now ASan-checks the
production native path and continuously guards these two fixes.


* chore: bump 0.0.1.49 (release #324+#325+#326+#327) + refresh stale ASan-lane comment

Integrator bump on the arena-alignment merge: triggers the first release since #163,
capturing the memory-debt fix (#324, VM->native gate, -40%), native codegen parity
(#325), the trait-vtable UB fix (#326), and this arena __int128 alignment + tk_mul_u16
fix (#327). Also refreshes the now-native ASan-lane comment (the step above is the
native gate, no longer a VM interpret).


* ci: LSan off on the native gate — arena holds process-lifetime allocs to exit (no per-test rewind)

The re-flipped native gate runs the whole corpus in one process with NO per-#test arena
rewind (that was a VM-test-runner mechanism), so the arena's process-lifetime allocations
(dsc_walk discovery strings, per-test crypto buffers) are all held to exit under the S0
discipline and reclaimed by the OS. LSan flagged ~132 KB of exactly those holds — NONE
from the alignment change (verified: stacks are teko__build__dsc_walk and
teko__crypto__sha256/512, not tk_chunk_alloc). LSan cannot tell a deliberate arena hold
from a real leak on a no-rewind run, so it is not a meaningful oracle on this lane;
detect_leaks=0 matches the sibling regressions-fixtures step. LSan's meaningful home is a
per-test-rewound run — the nightly VM lane.


---------

+++ src/runtime/teko_rt.h
@@ -108,8 +108,10 @@ TK_RT_LIST(tk_byte,   tk_byte_list)    // []byte  — byte builder, str of bytes
 TK_RT_LIST(tk_str,    tk_str_list)     // []str   — string accumulator lists (argv, paths, …)
 TK_RT_LIST(int64_t,   tk_i64_list)     // []i64   — integer accumulator lists
 
-// tk_alloc — the allocation seam (S0→S1). Hands back a fresh, uniquely-addressable,
-// max_align_t-aligned block of ≥ n usable bytes (n→1 when 0 so the result is unique);
+// tk_alloc — the allocation seam (S0→S1). Hands back a fresh, uniquely-addressable block of
+// ≥ n usable bytes (n→1 when 0 so the result is unique), aligned to the arena's alignment
+// (TK_ARENA_ALIGN in teko_rt.c: max of max_align_t's and __int128's, so __int128-carrying
+// Expr/TExpr nodes are correctly aligned even where max_align_t's alignment is only 8);
 // tk_panic on OOM (M.1, never NULL). Generated code allocates through this: slice
 // copy-append AND the auto-boxed recursive-value-type back-edges (tk_alloc(sizeof *p)).
 // (S1) The body now bump-allocates from the process ROOT region (tk_region_root) instead
@@ -122,9 +124,10 @@ TK_RT_LIST(int64_t,   tk_i64_list)     // []i64   — integer accumulator lists
 void *tk_alloc(size_t n);
 
 // ── Arena allocation (S1 — TEKO_EVOLUTION_DESIGN §5.2: arena primitive + root region) ──
-// A bump-allocator REGION: a chunk-list of malloc'd blocks, sub-allocated by a bump offset.
-// No per-object metadata, no free-list (M.0 metal/no-GC). region_alloc results are
-// max_align_t-aligned (malloc's own guarantee), so every type that was malloc-stored stays
+// A bump-allocator REGION: a chunk-list of aligned-malloc'd blocks, sub-allocated by a bump
+// offset. No per-object metadata, no free-list (M.0 metal/no-GC). region_alloc results are
+// TK_ARENA_ALIGN-aligned (an over-aligned chunk payload, NOT merely malloc's guarantee — see
+// teko_rt.c), so every type that was malloc-stored — including __int128-carrying nodes — stays
 // correctly aligned. OOM panics (M.1, never NULL). region_drop bulk-frees the whole span in
 // one pass (the S2 keystone). The process ROOT region (tk_region_root) is never dropped in
 // S1 → its memory lives for the whole process = today's malloc-everywhere leak (M.5
@@ -693,7 +696,7 @@ static inline int64_t  tk_sub_i64(int64_t  a, int64_t  b){ return a - b; }
 static inline __int128 tk_sub_i128(__int128 a, __int128 b){ return a - b; }
 
 static inline uint8_t  tk_mul_u8 (uint8_t  a, uint8_t  b){ return (uint8_t )(a * b); }
-static inline uint16_t tk_mul_u16(uint16_t a, uint16_t b){ return (uint16_t)(a * b); }
+static inline uint16_t tk_mul_u16(uint16_t a, uint16_t b){ return (uint16_t)((unsigned)a * (unsigned)b); }
 static inline uint32_t tk_mul_u32(uint32_t a, uint32_t b){ return a * b; }
 static inline uint64_t tk_mul_u64(uint64_t a, uint64_t b){ return a * b; }
 static inline unsigned __int128 tk_mul_u128(unsigned __int128 a, unsigned __int128 b){ return a * b; }
